Working port of pr-cli in Deno
This commit contains a fully functional port of my pr-cli written in go[^1]

It's using Deno[^2] with Cliffy[^3] for creating a command line tool. It uses
Gum[^4] for any interactive prompts.

This first version is functionally on-par with the Go version, but has an extra
external dependency: Gum. I'm hoping to solve this by bundling Gum in the binary.

There's also a lot of code cleanup to be done, especially around calling
external commands.

[^1]: https://github.com/annervisser/git-helper-scripts/tree/trunk/pr-cli
[^2]: https://deno.land
[^3]: https://cliffy.io
[^4]: https://github.com/charmbracelet/gum